+++ src/main/java/com/dailyon/paymentservice/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.dailyon.paymentservice.exception;
+package com.dailyon.paymentservice.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
