+++ src/main/java/com/dailyon/paymentservice/common/exception/advice/ApiControllerAdvice.java
@@ -1,7 +1,7 @@
-package com.dailyon.paymentservice.exception.advice;
+package com.dailyon.paymentservice.common.exception.advice;
 
-import com.dailyon.paymentservice.exception.CustomException;
-import com.dailyon.paymentservice.exception.response.ErrorResponse;
+import com.dailyon.paymentservice.common.exception.CustomException;
+import com.dailyon.paymentservice.common.exception.response.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +9,7 @@ import org.springframework.validation.BindException;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
